refactor calculations
