+++ packages/gallery/src/components/item/itemView.js
@@ -12,7 +12,7 @@ import {
 import ImageItem from './imageItem.js';
 import TextItem from './textItem.js';
 import ItemHover from './itemHover.js';
-import { changeActiveElementIfNeeded, onAnchorFocus } from './itemHelper.js';
+import { changeActiveElementIfNeeded, onAnchorFocus, getImageDimensions } from './itemHelper.js';
 import { cssScrollHelper } from '../helpers/cssScrollHelper';
 import { GalleryComponent } from '../galleryComponent';
 import {
@@ -344,35 +344,6 @@ class ItemView extends GalleryComponent {
 
   //---------------------------------------| COMPONENTS |-----------------------------------------//
 
-  getImageDimensions(itemProps) {
-    //image dimensions are for images in grid fit - placing the image with positive margins to show it within the square
-    const { styleParams, cubeRatio, style } = itemProps;
-    const { height: requiredHeight, width: requiredWidth, maxWidth, maxHeight, ratio } = style;
-    const { useMaxDimensions, itemBorderRadius, imageInfoType } = styleParams;
-    let dimensions = {};
-   
-    const _height = Math.min(requiredHeight, useMaxDimensions ? maxHeight : Infinity)
-    const _width = Math.min(requiredWidth, useMaxDimensions ? maxWidth : Infinity)
-    const imageMarginLeft = Math.round(
-      Math.max(0,( requiredWidth - _height * ratio) / 2)
-    );
-    const imageMarginTop = Math.round(
-      Math.max(0,( requiredHeight - _width / ratio) / 2)
-    );  
-    dimensions = {
-      height: requiredHeight - 2 * imageMarginTop,
-      width: requiredWidth - 2 * imageMarginLeft,
-      margin: `${imageMarginTop}px ${imageMarginLeft}px`,
-    };
-      
-    if (
-        itemBorderRadius &&
-        imageInfoType !== GALLERY_CONSTS.infoType.ATTACHED_BACKGROUND
-    ) {
-      dimensions.borderRadius = itemBorderRadius + 'px';
-    }
-    return dimensions;
-  }
 
   getItemHover(imageDimensions) {
     const { customHoverRenderer, ...props } = this.props;
@@ -478,7 +449,7 @@ class ItemView extends GalleryComponent {
   getItemInner() {
     const { styleParams, type } = this.props;
     let itemInner;
-    const imageDimensions = this.getImageDimensions(this.props);
+    const imageDimensions = getImageDimensions(this.props);
     const { width, height, margin } = imageDimensions;
 
     const itemStyles = { width, height };
@@ -778,7 +749,6 @@ class ItemView extends GalleryComponent {
       ...transitionStyles,
       ...opacityStyles,
       ...slideAnimationStyles,
-      border: '1px solid red'
     };
 
     return itemContainerStyles;
@@ -799,13 +769,12 @@ class ItemView extends GalleryComponent {
     styles.margin = -styleParams.itemBorderWidth + 'px';
     styles.height = height + 'px';
 
-    const imageDimensions = this.getImageDimensions(this.props);
+    const imageDimensions = getImageDimensions(this.props);
 
     const itemWrapperStyles = {
       ...styles,
       ...imageDimensions,
       ...(!styleParams.isSlideshow && this.getSlideAnimationStyles()),
-      border: '1px solid black'
     };
 
     return itemWrapperStyles;
